Update GOVERNANCE.md
This corrects the sentence stating that the PGB can take action under the CoC. The Code of Conduct, under enforcement responsibilities, states that OASIS staff are responsible for enforcing the CoC. This is to protect the PGB and OASIS from accusations of unfair practices. 
The proposed language is just a suggestion.